move fetch api to utils folder
