+++ apps/flow-docs/app/cms/utils/fetch-flips.tsx
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
-import { FlipCellProps } from "../../../../libs/design-system/src/lib/Components/FlipCell"
+import { FlipCellProps } from "../../../../../libs/design-system/src/lib/Components/FlipCell"
 import { Octokit } from "@octokit/core"
 
 export const fetchFlips = async () => {
